requirements.txt
google-genai
langchain
langchain-google-genai
pydantic
pillow
numpy